Makes current indentation logic more usable
The two symbols dart-indent-trigger-commands and
dart-indent-line-function are copied from python.el, as of commit:

<https://github.com/emacs-mirror/emacs/commit/05345babc988060cca540770599282102c34f2a7>

As we work to improve the smart indentation, this makes the current
state more usable, as the first TAB press uses the smart indentation,
and subsequent presses indent to tab stops, that is 2 spaces.
Additionally we bind S-TAB to dedent.

The function dart-indent-simple is inspired by other editors and IDEs.